# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: HaoHaoAccounting API
  version: 0.1.0
servers:
  - url: http://localhost:8080/api/v1
...
security:
  - bearerAuth: []
paths:
...
  /budgets:
    parameters:
      - $ref: '#/components/parameters/RequestID'
    get:
      tags: [budgets]
      operationId: getBudgets
      parameters:
        - name: month
          in: query
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
      responses:
        '200':
          description: Budget list
          headers:
            Vary:
              $ref: '#/components/headers/Vary'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Budget'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
    post:
      tags: [budgets]
      operationId: postBudgets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BudgetRequest'
      responses:
        '201':
          description: Created budget
          headers:
            Vary:
              $ref: '#/components/headers/Vary'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Budget'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
  /budgets/{id}:
    parameters:
      - $ref: '#/components/parameters/RequestID'
    put:
      tags: [budgets]
      operationId: putBudgetsById
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BudgetRequest'
      responses:
        '200':
          description: Updated budget
          headers:
            Vary:
              $ref: '#/components/headers/Vary'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Budget'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
    delete:
      tags: [budgets]
      operationId: deleteBudgetsById
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
  /categories:
    parameters:
      - $ref: '#/components/parameters/RequestID'
    get:
      tags: [categories]
      operationId: getCategories
      parameters:
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/TransactionType'
      responses:
        '200':
          description: Category list
          headers:
            Vary:
              $ref: '#/components/headers/Vary'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
    post:
      tags: [categories]
      operationId: postCategories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '201':
          description: Created category
          headers:
            Vary:
              $ref: '#/components/headers/Vary'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
  /categories/{id}:
    parameters:
      - $ref: '#/components/parameters/RequestID'
    put:
      tags: [categories]
      operationId: putCategoriesById
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '200':
          description: Updated category
          headers:
            Vary:
              $ref: '#/components/headers/Vary'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
    delete:
      tags: [categories]
      operationId: deleteCategoriesById
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    RequestID:
      name: X-Request-ID
      in: header
      required: false
      description: Optional request correlation id. The API returns this value when valid, otherwise it generates one.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[!-~]+$'
  headers:
    RequestID:
      description: Request correlation id for tracing logs and API calls.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[!-~]+$'
...
    Vary:
      description: Request headers that influenced response representation selection.
      schema:
        type: string
    Allow:
      description: HTTP methods supported by the target route.
      schema:
        type: string
    WWWAuthenticate:
      description: Authentication challenge for protected resources. Invalid bearer tokens include error="invalid_token" per RFC 6750.
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Bad request or validation error.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing, invalid, or expired authentication token.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
        WWW-Authenticate:
          $ref: '#/components/headers/WWWAuthenticate'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated user cannot perform this action.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource was not found.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    MethodNotAllowed:
      description: HTTP method is not supported for this route.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
        Allow:
          $ref: '#/components/headers/Allow'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    PayloadTooLarge:
      description: Request body is too large.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnsupportedMediaType:
      description: Request Content-Type is not supported by this operation.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotAcceptable:
      description: Requested response media type is not available for this operation.
      headers:
        Vary:
          $ref: '#/components/headers/Vary'
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    GatewayTimeout:
      description: Request processing exceeded the server timeout budget.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Ok:
      description: OK response
      headers:
        Vary:
          $ref: '#/components/headers/Vary'
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OkResponse'
  schemas:
    TransactionType:
      type: string
      enum: [income, expense]
...
    Budget:
      type: object
      required: [id, userId, month, categoryId, amount]
      properties:
        id:
          type: integer
        userId:
          type: integer
        month:
          type: string
        categoryId:
          type: integer
        amount:
          type: number
    BudgetRequest:
      type: object
      additionalProperties: false
      required: [month, amount]
      properties:
        month:
          type: string
          pattern: '^\d{4}-\d{2}$'
        categoryId:
          type: integer
          minimum: 1
        amount:
          type: number
          minimum: 0
    Category:
      type: object
      required: [id, name, type, isSystem]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        type:
          $ref: '#/components/schemas/TransactionType'
        isSystem:
          type: boolean
    CategoryRequest:
      type: object
      additionalProperties: false
      required: [name, type]
      properties:
        name:
          type: string
          minLength: 1
        type:
          $ref: '#/components/schemas/TransactionType'
...
    ErrorResponse:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
        code:
          type: string
          enum: [bad_request, invalid_request, unauthorized, forbidden, not_found, method_not_allowed, rate_limited, payload_too_large, unsupported_media_type, not_acceptable, request_timeout, client_closed_request, internal_error]
        requestId:
          type: string
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean